+++ source/Examples/Resources/DFSNamespaceFolder/1-DFSNamespaceFolder_Domain_MultipleTarget_Config.ps1
@@ -20,8 +20,9 @@
 <#
     .DESCRIPTION
         Create an AD Domain V2 based DFS namespace called software in the domain contoso.com with
-        a three targets on the servers ca-fileserver, ma-fileserver and ny-fileserver. It also
-        creates a IT folder in each namespace.
+        a four targets on the servers ca-fileserver, ma-fileserver, ny-fileserver01 and ny-filerserver02. It also
+        creates a IT folder in each namespace. The ny-fileserver02 IT folder target's state is configured to be offline.
+
 #>
 Configuration DFSNamespaceFolder_Domain_MultipleTarget_Config
 {
@@ -57,6 +58,7 @@ Configuration DFSNamespaceFolder_Domain_MultipleTarget_Config
         {
             Path                 = '\\contoso.com\software'
             TargetPath           = '\\ca-fileserver\software'
+            State                = 'Online'
             Ensure               = 'Present'
             Type                 = 'DomainV2'
             Description          = 'AD Domain based DFS namespace for storing software installers'
@@ -67,16 +69,29 @@ Configuration DFSNamespaceFolder_Domain_MultipleTarget_Config
         {
             Path                 = '\\contoso.com\software'
             TargetPath           = '\\ma-fileserver\software'
+            State                = 'Online'
             Ensure               = 'Present'
             Type                 = 'DomainV2'
             Description          = 'AD Domain based DFS namespace for storing software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceRoot Resource
 
-        DFSNamespaceRoot DFSNamespaceRoot_Domain_Software_NY
+        DFSNamespaceRoot DFSNamespaceRoot_Domain_Software_NY_01
         {
             Path                 = '\\contoso.com\software'
-            TargetPath           = '\\ny-fileserver\software'
+            TargetPath           = '\\ny-fileserver01\software'
+            State                = 'Online'
+            Ensure               = 'Present'
+            Type                 = 'DomainV2'
+            Description          = 'AD Domain based DFS namespace for storing software installers'
+            PsDscRunAsCredential = $Credential
+        } # End of DFSNamespaceRoot Resource
+
+        DFSNamespaceRoot DFSNamespaceRoot_Domain_Software_NY_02
+        {
+            Path                 = '\\contoso.com\software'
+            TargetPath           = '\\ny-fileserver02\software'
+            State                = 'Offline'
             Ensure               = 'Present'
             Type                 = 'DomainV2'
             Description          = 'AD Domain based DFS namespace for storing software installers'
@@ -89,6 +104,7 @@ Configuration DFSNamespaceFolder_Domain_MultipleTarget_Config
             Path                 = '\\contoso.com\software\it'
             TargetPath           = '\\ca-fileserver\it'
             Ensure               = 'Present'
+            State                = 'Online'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource
@@ -98,15 +114,27 @@ Configuration DFSNamespaceFolder_Domain_MultipleTarget_Config
             Path                 = '\\contoso.com\software\it'
             TargetPath           = '\\ma-fileserver\it'
             Ensure               = 'Present'
+            State                = 'Online'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource
 
-        DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY
+        DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY_01
         {
             Path                 = '\\contoso.com\software\it'
-            TargetPath           = '\\ny-fileserver\it'
+            TargetPath           = '\\ny-fileserver01\it'
             Ensure               = 'Present'
+            State                = 'Online'
+            Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
+            PsDscRunAsCredential = $Credential
+        } # End of DFSNamespaceFolder Resource
+
+        DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY_02
+        {
+            Path                 = '\\contoso.com\software\it'
+            TargetPath           = '\\ny-fileserver02\it'
+            Ensure               = 'Present'
+            State                = 'Offline'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource