+++ source/Examples/Resources/DFSNamespaceRoot/1-DFSNamespaceRoot_Domain_MultipleTarget_Config.ps1
@@ -20,8 +20,8 @@
 <#
     .DESCRIPTION
         Create an AD Domain V2 based DFS namespace called software in the domain contoso.com with
-        a three targets on the servers ca-fileserver, ma-fileserver and ny-fileserver. It also
-        creates a IT folder in each namespace.
+        a four targets on the servers ca-fileserver, ma-fileserver, ny-fileserver01 and ny-filerserver02. It also
+        creates a IT folder in each namespace. The ny-fileserver02 IT folder target's state is configured to be offline.
 #>
 Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
 {
@@ -89,6 +89,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
             Path                 = '\\contoso.com\software\it'
             TargetPath           = '\\ca-fileserver\it'
             Ensure               = 'Present'
+            State                = 'Online'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource
@@ -98,15 +99,27 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
             Path                 = '\\contoso.com\software\it'
             TargetPath           = '\\ma-fileserver\it'
             Ensure               = 'Present'
+            State                = 'Online'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource
 
-        DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY
+        DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY_01
         {
             Path                 = '\\contoso.com\software\it'
-            TargetPath           = '\\ny-fileserver\it'
+            TargetPath           = '\\ny-fileserver01\it'
             Ensure               = 'Present'
+            State                = 'Online'
+            Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
+            PsDscRunAsCredential = $Credential
+        } # End of DFSNamespaceFolder Resource
+
+        DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY02
+        {
+            Path                 = '\\contoso.com\software\it'
+            TargetPath           = '\\ny-fileserver02\it'
+            Ensure               = 'Present'
+            State                = 'Online'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource